+++ test/resolver_test.dart
@@ -10,6 +10,13 @@ import 'package:test_descriptor/test_descriptor.dart' as d;
 void main() {
   group('Default Resolver', () {
     setUp(() async {
+      final String sandboxUriPath = p.toUri(d.sandbox).toString();
+      await d.dir('bar', [
+        d.dir('lib', [
+          d.file('bar.dart', 'final fizz = "bar";'),
+        ])
+      ]).create();
+
       await d.dir('foo', [
         d.dir('.dart_tool', [
           d.file('bad_package_config.json', 'thisIsntAPackageConfigFile!'),
@@ -19,7 +26,12 @@ void main() {
   "packages": [
     {
       "name": "foo",
-      "rootUri": "file:///${d.sandbox}/foo",
+      "rootUri": "../",
+      "packageUri": "lib/"
+    },
+    {
+      "name": "bar",
+      "rootUri": "$sandboxUriPath/bar",
       "packageUri": "lib/"
     }
   ]
@@ -37,14 +49,16 @@ void main() {
           packagesPath:
               p.join(d.sandbox, 'foo', '.dart_tool', 'package_config.json'));
       expect(resolver.resolve('package:foo/foo.dart'),
-          '${d.sandbox}/foo/lib/foo.dart');
+          p.join(d.sandbox, 'foo', 'lib', 'foo.dart'));
+      expect(resolver.resolve('package:bar/bar.dart'),
+          p.join(d.sandbox, 'bar', 'lib', 'bar.dart'));
     });
 
     test('can be created from a package directory', () async {
       final resolver =
           await Resolver.create(packagePath: p.join(d.sandbox, 'foo'));
       expect(resolver.resolve('package:foo/foo.dart'),
-          '${d.sandbox}/foo/lib/foo.dart');
+          p.join(d.sandbox, 'foo', 'lib', 'foo.dart'));
     });
 
     test('errors if the packagesFile is an unknown format', () async {